+++ ecs/scripts/generate-and-push-templates.sh
@@ -7,6 +7,8 @@ TEMPLATES_PATH="plugins/setup/src/main/resources/"
 CURRENT_PATH=$(pwd)
 OUTPUT_PATH=${OUTPUT_PATH:-"$CURRENT_PATH"/../output}
 BASE_BRANCH=${BASE_BRANCH:-main}
+DOCUMENTATION_PATH="docs"
+CSV_SUBPATH="mappings/${ECS_VERSION}/generated/csv/fields.csv"
 
 # Committer's identity
 COMMITER_EMAIL=${COMMITER_EMAIL:-$(git config user.email)}
@@ -53,13 +55,14 @@ detect_modified_modules() {
     modified_files=$(git diff --name-only origin/"$BASE_BRANCH")
 
     for file in $modified_files; do
-        if [[ $file == ecs/* ]]; then
+        if [[ $file == ecs/* && ( $file == *.yml || $file == *.json ) ]]; then
             ecs_module=$(echo "$file" | cut -d'/' -f2)
             if [[ ! " ${updated_modules[*]} " =~ ${ecs_module} ]]; then
                 updated_modules+=("$ecs_module")
             fi
         fi
     done
+
     echo "Updated ECS modules: ${updated_modules[*]}"
 
     # Mapping section
@@ -77,6 +80,7 @@ detect_modified_modules() {
         [states-inventory-scheduled-commands]="index-template-scheduled-commands.json"
         [states-inventory-system]="index-template-system.json"
         [states-vulnerabilities]="index-template-vulnerabilities.json"
+        [users]="index-template-users.json"
     )
 
     relevant_modules=()
@@ -118,6 +122,7 @@ configure_git() {
     chmod 644 ~/.ssh/id_ed25519_bot.pub
 
     # Setup commit signing
+    eval "$(ssh-agent -s)"
     ssh-add ~/.ssh/id_ed25519_bot
     git config --global gpg.format ssh
     git config --global commit.gpgsign true
@@ -135,7 +140,7 @@ commit_and_push_changes() {
     echo
     echo "---> Committing and pushing changes to ${PLUGINS_REPO} repository..."
 
-    echo "Copying ECS templates to the plugins repository..."
+    echo "Copying ECS templates and csv definitions to the plugins repository..."
     for ecs_module in "${relevant_modules[@]}"; do
         target_file=${module_to_file[$ecs_module]}
         if [[ -z "$target_file" ]]; then
@@ -146,7 +151,14 @@ commit_and_push_changes() {
         cp "$CURRENT_PATH/ecs/$ecs_module/$MAPPINGS_SUBPATH" "$OUTPUT_PATH/$target_file"
         # Copy the template to the plugins repository
         echo "  - Copy template for module '$ecs_module' to '$target_file'"
+        # If the target file does not exist, create it.
+        if [ ! -f "$TEMPLATES_PATH/$target_file" ]; then
+            touch "$TEMPLATES_PATH/$target_file"
+        fi
         cp "$CURRENT_PATH/ecs/$ecs_module/$MAPPINGS_SUBPATH" "$TEMPLATES_PATH/$target_file"
+        # Copy the csv to the plugins repository
+        echo "  - Copy the updated csv definitions for module '$ecs_module' to '$CURRENT_PATH/ecs/$ecs_module/$DOCUMENTATION_PATH'"
+        cp "$CURRENT_PATH/ecs/$ecs_module/$CSV_SUBPATH" "$CURRENT_PATH/ecs/$ecs_module/$DOCUMENTATION_PATH/"
     done
 
     git status --short
@@ -250,7 +262,6 @@ main() {
     validate_dependencies
     detect_modified_modules
     run_ecs_generator # Exit if no changes on relevant modules.
-    clone_target_repo
     commit_and_push_changes # Exit if no changes detected.
     create_or_update_pr
 }