+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/util/Privileged.java
@@ -16,8 +16,6 @@
  */
 package com.wazuh.contentmanager.util;
 
-import org.apache.hc.client5.http.async.methods.SimpleHttpResponse;
-
 import java.security.AccessController;
 import java.security.PrivilegedAction;
 
@@ -27,11 +25,10 @@ public class Privileged {
     /**
      * Executes an HTTP request with elevated privileges.
      *
-     * @param request A privileged action that performs the HTTP request.
-     * @return The SimpleHttpResponse resulting from the request execution.
+     * @param <T> A privileged action that performs the HTTP request.
+     * @return The return value resulting from the request execution.
      */
-    public static SimpleHttpResponse doPrivilegedRequest(
-            PrivilegedAction<SimpleHttpResponse> request) {
+    public static <T> T doPrivilegedRequest(java.security.PrivilegedAction<T> request) {
         return AccessController.doPrivileged(request);
     }
 }